feat: [PIE-2530]: Creation of generic stepNode plan creator
